Ignore the easyssh.go and util/*.go for test coverage

+++ test-coverage.sh
@@ -6,16 +6,26 @@
 # go get github.com/mattn/goveralls
 # COVERALLS="-service drone.io -repotoken $COVERALLS_TOKEN" ./test-coverage.sh
 #
+# NOTE: go files in the root and in the dir "utils" are ignored for purposes of
+#       test coverage calculation.
+
+set -eo pipefail
 
 echo "mode: count" > acc.out
 fail=0
+
+want-coverage-for() {
+    if [[ $1 == . || $1 == ./util ]]; then
+        return 1
+    fi
+}
 
 # Standard go tooling behavior is to ignore dirs with leading underscors
 for dir in $(find . -maxdepth 10 -not -path './.git*' -not -path '*/_*' -type d);
 do
   if ls $dir/*.go &> /dev/null; then
     go test -covermode=count -coverprofile=profile.out $dir || fail=1
-    if [ -f profile.out ]
+    if [ -f profile.out ] && want-coverage-for $dir
     then
       cat profile.out | grep -v "^mode: " | grep -v "test_helpers.go" >> acc.out
       rm profile.out
